pyproject.toml
[project]
name = "mcp-server-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "autogen-ext",
    "autogen-agentchat",
    "autogen-core",
    "python-dotenv",
    "openai",
    "tiktoken",
    "json_schema_to_pydantic",
    "aiohttp>=3.12.15",
    "pyjwt>=2.10.1",
    "fastmcp>=2.12.3",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-api-python-client>=2.183.0",
    "pymongo>=4.15.2",
    "numpy>=2.3.3",
]

